+++ client/src/lib/docs/index.js
@@ -1,4 +1,4 @@
-const docs = require("./docs.json");
+const docs = { ...require("./docs.json") };
 const releaseNotes = require("./release-notes.json");
 
 module.exports = {