+++ CHANGELOG.md
@@ -21,6 +21,8 @@ All notable changes to this project will be documented in this file.
 - Fixed an issues causing the Power VS Instance form to not be open on load with an invalid configuration
 - Fixed an issue causing Power VS workspace form to crash when entering a name with no other fields selected
 - Fixed an issue causing options page to incorrectly unset craig version on update
+- Fixed an issue causing the template setup wizard to not persist High Availability for Power VS
+- Fixed an issue causing Docs to have invalid tables when converted to markdown
 
 ## 1.3.0